+ fetch data + add tasks to coredata

+++ Viper_SwiftUI_CoreData/EntityModels/TaskJSON.swift
@@ -12,7 +12,7 @@ struct Todos: Decodable {
 }
 
 struct Todo: Decodable {
-    var id,userId: Int
+    var id,userId: Double
     var completed: Bool
     var todo: String
